load-service: Validate capabilities
This is useful for detecting invalid capabilties in dinitcheck.
