+++ src/includes/dinit-util.h
@@ -16,6 +16,11 @@
 #include <unistd.h>
 
 #include <baseproc-sys.h>
+#include <mconfig.h>
+
+#if SUPPORT_CAPABILITIES
+#include <sys/capability.h>
+#endif
 
 // Check if a value is one of several possible values.
 // Use like:     value(x).is_in(1,2,3)
@@ -814,5 +819,42 @@ public:
         return std::string(str_data + pos, sub_len);
     }
 };
+
+#if SUPPORT_CAPABILITIES
+// A thin wrapper around the structure to avoid having to worry about freeing
+// it in various places (move semantics and destructor will take care of it)
+struct cap_iab_wrapper {
+    cap_iab_wrapper() {}
+    cap_iab_wrapper(std::string const &str) noexcept {
+        if (str.empty()) return;
+        // this may end up being nullptr
+        // throwing from constructors is bad, so always check .get() afterwards
+        iab = cap_iab_from_text(str.c_str());
+    }
+
+    cap_iab_wrapper(cap_iab_wrapper const &) = delete;
+    cap_iab_wrapper(cap_iab_wrapper &&v) noexcept: iab(v.iab) {
+        v.iab = nullptr;
+    }
+
+    cap_iab_wrapper &operator=(cap_iab_wrapper const &) = delete;
+    cap_iab_wrapper &operator=(cap_iab_wrapper &&v) noexcept {
+        iab = v.iab;
+        v.iab = nullptr;
+        return *this;
+    }
+
+    ~cap_iab_wrapper() noexcept {
+        if (iab) cap_free(iab);
+    }
+
+    cap_iab_t get() const noexcept {
+        return iab;
+    }
+
+private:
+    cap_iab_t iab = nullptr;
+};
+#endif
 
 #endif
