+++ src/includes/proc-service.h
@@ -12,48 +12,12 @@
 #include <baseproc-sys.h>
 #include <service.h>
 #include <dinit-utmp.h>
+#include <dinit-util.h>
 
 // This header defines base_proc_service (base process service) and several derivatives, as well as some
 // utility functions and classes. See service.h for full details of services.
 
 class process_service;
-
-#if SUPPORT_CAPABILITIES
-// A thin wrapper around the structure to avoid having to worry about freeing
-// it in various places (move semantics and destructor will take care of it)
-struct cap_iab_wrapper {
-    cap_iab_wrapper() {}
-    cap_iab_wrapper(std::string const &str) noexcept {
-        if (str.empty()) return;
-        // this may end up being nullptr
-        // throwing from constructors is bad, so always check .get() afterwards
-        iab = cap_iab_from_text(str.c_str());
-    }
-
-    cap_iab_wrapper(cap_iab_wrapper const &) = delete;
-    cap_iab_wrapper(cap_iab_wrapper &&v) noexcept: iab(v.iab) {
-        v.iab = nullptr;
-    }
-
-    cap_iab_wrapper &operator=(cap_iab_wrapper const &) = delete;
-    cap_iab_wrapper &operator=(cap_iab_wrapper &&v) noexcept {
-        iab = v.iab;
-        v.iab = nullptr;
-        return *this;
-    }
-
-    ~cap_iab_wrapper() noexcept {
-        if (iab) cap_free(iab);
-    }
-
-    cap_iab_t get() const noexcept {
-        return iab;
-    }
-
-private:
-    cap_iab_t iab = nullptr;
-};
-#endif
 
 // Given a string and a list of pairs of (start,end) indices for each argument in that string,
 // store a null terminator for the argument. Return a `char *` vector containing the beginning
